chore: moved KeyStore back to its own file
